+++ core/wait/waiterhelper.go
@@ -44,7 +44,7 @@ func (w *WaiterHelper[T, S]) Wait() AsyncActionCheck[T] {
 					return true, nil, nil
 				}
 			}
-			return true, nil, err
+			return false, nil, err
 		}
 
 		state, err := w.GetState(instance)